move stack to history manager
